heyongqiang sixth submit

+++ project-buyer/src/main/java/com/heyongqiang/work/dao/pojo/Passenger.java
@@ -16,7 +16,7 @@ public class Passenger {
 
     private String telephone;
 
-    private Integer certificateId;
+    private Long certificateId;
 
     private String email;
